feat: Extract media source capabilities and implement it in other files
