+++ AUTHORS
@@ -94,3 +94,4 @@ Raymond Cheng <raycheng100@gmail.com>
 Blue Billywig <*@bluebillywig.com>
 João Nabais <jlnabais@gmail.com>
 Koen Romers <koenromers@gmail.com>
+Zhenghang Chen <czhtju@gmail.com>
