+++ lib/media/media_source_engine.js
@@ -8,6 +8,7 @@ goog.provide('shaka.media.MediaSourceEngine');
 
 goog.require('goog.asserts');
 goog.require('shaka.log');
+goog.require('shaka.media.Capabilities');
 goog.require('shaka.media.ContentWorkarounds');
 goog.require('shaka.media.ClosedCaptionParser');
 goog.require('shaka.media.IClosedCaptionParser');
@@ -175,24 +176,8 @@ shaka.media.MediaSourceEngine = class {
         stream.mimeType, stream.codecs);
     const extendedMimeType = shaka.util.MimeUtils.getExtendedType(stream);
     return shaka.text.TextEngine.isTypeSupported(fullMimeType) ||
-        shaka.media.MediaSourceEngine.isTypeSupported(extendedMimeType) ||
+        shaka.media.Capabilities.isTypeSupported(extendedMimeType) ||
         shaka.media.Transmuxer.isSupported(fullMimeType, stream.type);
-  }
-
-  /**
-   * Cache browser engine call to improve performance on some poor platforms
-   *
-   * @param {string} type
-   * @return {boolean}
-   */
-  static isTypeSupported(type) {
-    const supportMap = shaka.media.MediaSourceEngine.MediaSourceTypeSupportMap_;
-    if (supportMap.has(type)) {
-      return supportMap.get(type);
-    }
-    const currentSupport = MediaSource.isTypeSupported(type);
-    supportMap.set(type, currentSupport);
-    return currentSupport;
   }
 
   /**
@@ -247,7 +232,7 @@ shaka.media.MediaSourceEngine = class {
         if (shaka.text.TextEngine.isTypeSupported(type)) {
           support[type] = true;
         } else {
-          support[type] = shaka.media.MediaSourceEngine.isTypeSupported(type) ||
+          support[type] = shaka.media.Capabilities.isTypeSupported(type) ||
                           shaka.media.Transmuxer.isSupported(type);
         }
       } else {
@@ -382,7 +367,7 @@ shaka.media.MediaSourceEngine = class {
         this.reinitText(mimeType, sequenceMode);
       } else {
         if ((forceTransmux ||
-            !shaka.media.MediaSourceEngine.isTypeSupported(mimeType)) &&
+            !shaka.media.Capabilities.isTypeSupported(mimeType)) &&
             shaka.media.Transmuxer.isSupported(mimeType, contentType)) {
           this.transmuxers_[contentType] =
               new shaka.media.Transmuxer(mimeType);
@@ -1406,8 +1391,3 @@ shaka.media.MediaSourceEngine.RAW_FORMATS = [
   'audio/ec3',
   'audio/mpeg',
 ];
-
-/**
-  * @private {!Map.<string, boolean>}
-*/
-shaka.media.MediaSourceEngine.MediaSourceTypeSupportMap_ = new Map();
