+++ lib/media/transmuxer.js
@@ -7,6 +7,7 @@
 goog.provide('shaka.media.Transmuxer');
 
 goog.require('goog.asserts');
+goog.require('shaka.media.Capabilities');
 goog.require('shaka.util.BufferUtils');
 goog.require('shaka.util.Error');
 goog.require('shaka.util.IDestroyable');
@@ -71,6 +72,7 @@ shaka.media.Transmuxer = class {
    */
   static isSupported(mimeType, contentType) {
     const Transmuxer = shaka.media.Transmuxer;
+    const Capabilities = shaka.media.Capabilities;
 
     const isTs = Transmuxer.isTsContainer_(mimeType);
     const isAac = Transmuxer.isAacContainer_(mimeType);
@@ -80,11 +82,11 @@ shaka.media.Transmuxer = class {
     }
 
     if (isAac) {
-      return MediaSource.isTypeSupported(Transmuxer.convertAacCodecs_());
+      return Capabilities.isTypeSupported(Transmuxer.convertAacCodecs_());
     }
 
     if (contentType) {
-      return MediaSource.isTypeSupported(
+      return Capabilities.isTypeSupported(
           Transmuxer.convertTsCodecs_(contentType, mimeType));
     }
 
@@ -92,8 +94,8 @@ shaka.media.Transmuxer = class {
 
     const audioMime = Transmuxer.convertTsCodecs_(ContentType.AUDIO, mimeType);
     const videoMime = Transmuxer.convertTsCodecs_(ContentType.VIDEO, mimeType);
-    return MediaSource.isTypeSupported(audioMime) ||
-        MediaSource.isTypeSupported(videoMime);
+    return Capabilities.isTypeSupported(audioMime) ||
+        Capabilities.isTypeSupported(videoMime);
   }
