+++ lib/polyfill/media_capabilities.js
@@ -7,6 +7,7 @@
 goog.provide('shaka.polyfill.MediaCapabilities');
 
 goog.require('shaka.log');
+goog.require('shaka.media.Capabilities');
 goog.require('shaka.polyfill');
 goog.require('shaka.util.Error');
 goog.require('shaka.util.Platform');
@@ -100,11 +101,14 @@ shaka.polyfill.MediaCapabilities = class {
     const videoConfig = mediaDecodingConfig['video'];
     const audioConfig = mediaDecodingConfig['audio'];
 
+    const Capabilities = shaka.media.Capabilities;
+
     if (mediaDecodingConfig.type == 'media-source') {
       if (!shaka.util.Platform.supportsMediaSource()) {
         return res;
       }
-      // Use 'MediaSource.isTypeSupported' to check if the stream is supported.
+      // Use 'shaka.media.Capabilities.isTypeSupported'to check if
+      // the stream is supported.
       // Cast platforms will additionally check canDisplayType(), which
       // accepts extended MIME type parameters.
       // See: https://github.com/shaka-project/shaka-player/issues/4726
@@ -114,7 +118,7 @@ shaka.polyfill.MediaCapabilities = class {
           isSupported =
               shaka.polyfill.MediaCapabilities.canCastDisplayType_(videoConfig);
         } else {
-          isSupported = MediaSource.isTypeSupported(videoConfig.contentType);
+          isSupported = Capabilities.isTypeSupported(videoConfig.contentType);
         }
         if (!isSupported) {
           return res;
@@ -123,7 +127,7 @@ shaka.polyfill.MediaCapabilities = class {
 
       if (audioConfig) {
         const contentType = audioConfig.contentType;
-        const isSupported = MediaSource.isTypeSupported(contentType);
+        const isSupported = Capabilities.isTypeSupported(contentType);
         if (!isSupported) {
           return res;
         }
@@ -246,8 +250,8 @@ shaka.polyfill.MediaCapabilities = class {
           shaka.util.Error.Code.CAST_API_UNAVAILABLE);
     } else if (!(cast.__platform__ && cast.__platform__.canDisplayType)) {
       shaka.log.warning('Expected cast APIs to be available! Falling back to ' +
-          'MediaSource.isTypeSupported() for type support.');
-      return MediaSource.isTypeSupported(videoConfig.contentType);
+          'shaka.media.Capabilities.isTypeSupported() for type support.');
+      return shaka.media.Capabilities.isTypeSupported(videoConfig.contentType);
     }
 
     let displayType = videoConfig.contentType;
