+++ lib/util/stream_utils.js
@@ -8,7 +8,7 @@ goog.provide('shaka.util.StreamUtils');
 
 goog.require('goog.asserts');
 goog.require('shaka.log');
-goog.require('shaka.media.MediaSourceEngine');
+goog.require('shaka.media.Capabilities');
 goog.require('shaka.text.TextEngine');
 goog.require('shaka.util.Functional');
 goog.require('shaka.util.LanguageUtils');
@@ -438,6 +438,7 @@ shaka.util.StreamUtils = class {
       // See: https://github.com/shaka-project/shaka-player/issues/3860
       const video = variant.video;
       const ContentType = shaka.util.ManifestParserUtils.ContentType;
+      const Capabilities = shaka.media.Capabilities;
       if (video) {
         let videoCodecs =
             shaka.util.StreamUtils.getCorrectVideoCodecs_(video.codecs);
@@ -455,13 +456,13 @@ shaka.util.StreamUtils = class {
               shaka.util.StreamUtils.getCorrectAudioCodecs_(audioCodecs);
           const audioFullType = shaka.util.MimeUtils.getFullOrConvertedType(
               video.mimeType, audioCodecs, ContentType.AUDIO);
-          if (!shaka.media.MediaSourceEngine.isTypeSupported(audioFullType)) {
+          if (!Capabilities.isTypeSupported(audioFullType)) {
             return false;
           }
         }
         const fullType = shaka.util.MimeUtils.getFullOrConvertedType(
             video.mimeType, videoCodecs, ContentType.VIDEO);
-        if (!shaka.media.MediaSourceEngine.isTypeSupported(fullType)) {
+        if (!Capabilities.isTypeSupported(fullType)) {
           return false;
         }
       }
@@ -471,7 +472,7 @@ shaka.util.StreamUtils = class {
             shaka.util.StreamUtils.getCorrectAudioCodecs_(audio.codecs);
         const fullType = shaka.util.MimeUtils.getFullOrConvertedType(
             audio.mimeType, codecs, ContentType.AUDIO);
-        if (!shaka.media.MediaSourceEngine.isTypeSupported(fullType)) {
+        if (!Capabilities.isTypeSupported(fullType)) {
           return false;
         }
       }
